add components to client form, add, update, delete, code on the buttons
added multiple buttons added a tab control for add, update and delete. Added data grids and buttons and textboxes on each of the three tabs coded the add button to add a client to the database, coded the clear button to clear all the tex boxes. On the other update tab coded on the update button to make changes to the existing database if there was a problem with the information and lastly the delete tab coded in the delete button to delete a specific client ID from the database